# internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Printa Ecosystem API
  version: 1.0.0
  description: |
    The Printa Ecosystem API supports vendor onboarding, stores, catalogues, orders,
    routing, production, point of sale, billing, payments, notifications, and external
    communications. All timestamps are RFC 3339 UTC strings and all identifiers are UUIDs
    unless an endpoint documents otherwise.

    Protected endpoints require an `Authorization: Bearer <JWT>` header. Roles are enforced
    by the backend; an operation description identifies any additional role restriction.
  contact:
    name: Printa Engineering
servers:
  - url: /
    description: Current environment
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/billing/invoices/{id}/pay:
    parameters: [ { $ref: '#/components/parameters/ID' } ]
    post:
      tags: [Billing]
      summary: Mark an invoice as paid
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarkInvoicePaid' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/webhooks/mtn-momo:
    post:
      tags: [Payments]
      summary: Receive an MTN MoMo payment callback
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentWebhook' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/webhooks/airtel-money:
    post:
      tags: [Payments]
      summary: Receive an Airtel Money payment callback
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentWebhook' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/payments:
    post:
      tags: [Payments]
      summary: Initiate a payment
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InitiatePayment' }
      responses:
        '201': { $ref: '#/components/responses/ObjectCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/payments/{id}:
    parameters: [ { $ref: '#/components/parameters/ID' } ]
    get:
      tags: [Payments]
      summary: Get a payment transaction
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/payments/{id}/verify:
    parameters: [ { $ref: '#/components/parameters/ID' } ]
    post:
      tags: [Payments]
      summary: Verify a payment with its provider
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/payments/{id}/refund:
    parameters: [ { $ref: '#/components/parameters/ID' } ]
    post:
      tags: [Payments]
      summary: Refund a payment transaction
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Refund' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/payments/reference/{ref_type}/{ref_id}:
    parameters:
      - name: ref_type
        in: path
        required: true
        schema: { type: string, enum: [ORDER, INVOICE, SUBSCRIPTION] }
      - name: ref_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Payments]
      summary: List payments by business reference
      responses:
        '200': { $ref: '#/components/responses/ArrayResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/payments/vendor/{vendor_id}:
    parameters: [ { $ref: '#/components/parameters/VendorID' } ]
    get:
      tags: [Payments]
      summary: List payments for a vendor
      responses:
        '200': { $ref: '#/components/responses/ArrayResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    VendorID:
      name: vendor_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: A caller-generated key used to safely retry an idempotent operation.
      schema: { type: string, maxLength: 255 }
  responses:
    ObjectResponse:
      description: Successful object response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenericObject' }
    ArrayResponse:
      description: Successful collection response
      content:
        application/json:
          schema:
            type: array
            items: { $ref: '#/components/schemas/GenericObject' }
    ObjectCreated:
      description: Resource created
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenericObject' }
    BadRequest:
      description: Invalid request body, query parameter, or state transition
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid, expired, or inactive JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    GenericObject:
      type: object
      additionalProperties: true
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Refund:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
...
    MarkInvoicePaid:
      type: object
      required: [payment_reference]
      properties:
        payment_reference: { type: string }
        notes: { type: string }
    InitiatePayment:
      type: object
      required: [provider, reference_type, reference_id, amount]
      properties:
        provider: { type: string, enum: [MTN_MOMO, AIRTEL_MONEY, CASH, CARD] }
        reference_type: { type: string, enum: [ORDER, INVOICE, SUBSCRIPTION] }
        reference_id: { type: string, format: uuid }
        vendor_id: { type: string, format: uuid }
        amount: { type: number, format: double, minimum: 0 }
        currency: { type: string, default: ZMW }
        phone_number: { type: string }
        description: { type: string }
        idempotency_key: { type: string }
    PaymentWebhook:
      type: object
      required: [provider, external_ref, status, amount, currency, raw_payload]
      properties:
        provider: { type: string }
        external_ref: { type: string }
        status: { type: string }
        amount: { type: number, format: double }
        currency: { type: string }
        phone_number: { type: string }
        raw_payload:
          type: object
          additionalProperties: true
...